refactor of the sample converters and transformers
the reference and fast sample converters are now exposed via the basic
sample transformer (and the batch version). this allows us to also
expose reference and fast sample transformers as they can be expressed
in terms of the sample converters. the convert and transform code is
now templated on sample converter and sample transformer respectively.
this allows us to create reference_convert and fast_convert functions
to sit alongside convert (which now uses the default sample converter)
and reference_transform and fast_transform functions to sit alongside
transform (which now uses the default sample transformer, which uses the
default sample converter).