make containers allocator-aware
the containers are currently not allocater aware, i.e. they don't
pointer types that are not sample_type*. This changes adds support for
fancy pointers in all the containers and also propagates those types
into any output spans or interators.

Also changes last of the uses of class in the template params to
typename (except in the case of template template parameters)