use sfinae to remove need for float to float specialisation
prior to this change we needed a dummy (i.e. returns input) float to
float specialisation otherwise it would try to use the X to float or
float to X specialisations, which was not only incorrect but led to
ambiguous specialisation error. By disabling the X to float and float to
X overloads if X is not an integer then we no longer need the dummy
float to float specialisation