Add chrono
sample_duration, sample_time_point, subsample_duration, and
subsample_time_point and probably uncontroversial, but I'm a little
unsure about the maths in forward_delay_locked_loop and
reverse_delay_locked_loop. I'm going to get someone who's better at math
than I am to look it over and validate the correctness of the approach.
Both the DLL implementations are based off of
https://kokkinizita.linuxaudio.org/papers/usingdll.pdf, though I've had
to extend it slightly in order to add the functionality to map a period
of time from one clock domain to another.