Create generalised underlying_cast
narrowing_cast doesn't work on network-byte-ordered values and hence the
assert in narrowing_cast is triggered when used there. To solve this I
created a new value_cast function that is able to cast between both
larger and smaller types, then reimplemented narrowing_cast in terms of
this (renaming it to sample_value_narrowing_cast) then using the plain
value_cast for the network-byte-order casts.