use sample_iterator to pass types to be blitted in transformer
Prior to this change I passed a buffer of samples to be blitted via a
pointer. This change makes it so that we now pass the buffers via a
sample_iterator. This simplifies a lot of logic in the batch creator as
all input and output iterators are now sample_iterators.